Some refactoring of azappconfig
Moved options and response types to their respective files.
Moved methods on Setting to its respective file.
Inlined various from* conversion funcs to their call sites.
Renamed paged response types to have the Response suffix per guidelines.
Fixed an issue parsing the time value returned in the Last-Modified
header. The expected format was incorrect and the error was previously
silently dropped.
