+++ sdk/data/azappconfig/setting.go
@@ -83,3 +83,39 @@ func (cs Setting) toGenerated() generated.KeyValue {
 		Value:        cs.Value,
 	}
 }
+
+func (cs Setting) toGeneratedDeleteLockOptions(ifMatch *azcore.ETag) *generated.AzureAppConfigurationClientDeleteLockOptions {
+	return &generated.AzureAppConfigurationClientDeleteLockOptions{
+		IfMatch: toGeneratedETagString(ifMatch),
+		Label:   cs.Label,
+	}
+}
+
+func (cs Setting) toGeneratedGetOptions(ifNoneMatch *azcore.ETag, acceptDateTime *time.Time) *generated.AzureAppConfigurationClientGetKeyValueOptions {
+	var dt *string
+	if acceptDateTime != nil {
+		str := acceptDateTime.Format(timeFormat)
+		dt = &str
+	}
+
+	return &generated.AzureAppConfigurationClientGetKeyValueOptions{
+		AcceptDatetime: dt,
+		IfNoneMatch:    toGeneratedETagString(ifNoneMatch),
+		Label:          cs.Label,
+	}
+}
+
+func (cs Setting) toGeneratedPutLockOptions(ifMatch *azcore.ETag) *generated.AzureAppConfigurationClientPutLockOptions {
+	return &generated.AzureAppConfigurationClientPutLockOptions{
+		IfMatch: toGeneratedETagString(ifMatch),
+		Label:   cs.Label,
+	}
+}
+
+func (cs Setting) toGeneratedPutOptions(ifMatch *azcore.ETag, ifNoneMatch *azcore.ETag) (generated.KeyValue, generated.AzureAppConfigurationClientPutKeyValueOptions) {
+	return cs.toGenerated(), generated.AzureAppConfigurationClientPutKeyValueOptions{
+		IfMatch:     toGeneratedETagString(ifMatch),
+		IfNoneMatch: toGeneratedETagString(ifNoneMatch),
+		Label:       cs.Label,
+	}
+}
